Rename defaultStyles to DefaultStyles
