+++ app/(tabs)/Logout.tsx
@@ -2,7 +2,7 @@ import { useRouter } from "expo-router";
 import { View, TouchableOpacity, Text } from "react-native";
 import { useAuth0 } from "react-native-auth0";
 
-import { defaultStyles } from "@/stylesheets/defaultStyles";
+import { defaultStyles } from "@/stylesheets/DefaultStyles";
 
 export default function Logout() {
   const router = useRouter();
